# src/main/resources/openapi/banco.yaml
openapi: 3.0.1
info:
  title: API Banco
  version: 1.0.0

paths:
  /api/banco/cliente/{codCliente}:
    get:
      summary: Obtener información de cliente
      parameters:
        - name: codCliente
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Información del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteCuenta'

  /api/banco/empleado:
    post:
      summary: Registrar un empleado
      operationId: registrarEmpleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Empleado'
      responses:
        "200":
          description: Empleado registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empleado'

  /api/banco/empleados:
    get:
      summary: Listar empleados
      operationId: listarEmpleados
      responses:
        "200":
          description: Lista de empleados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Empleado'

  /api/banco/empleado/{dni}:
    get:
      summary: Buscar empleado por DNI
      operationId: buscarEmpleadoPorDni
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empleado'

    delete:
      summary: Eliminar empleado
      operationId: eliminarEmpleado
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Empleado eliminado correctamente

components:
  schemas:

    Empleado:
      type: object
      properties:
        dni:
          type: string
          example: "87654321"
        nombre:
          type: string
          example: "Carlos"
        apellidos:
          type: string
          example: "Ramirez"
        puesto:
          type: string
          example: "Cajero"
        edad:
          type: integer
          example: 28

    ClienteCuenta:
      type: object
      properties:
        codCliente:
          type: string
          example: "C001"
        nombre:
          type: string
          example: "Juan Perez"
        saldo:
          type: number
          example: 1500.50